% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bop}
\alias{bop}
\alias{bop.docs}
\alias{bop.docs.metadata}
\alias{bop.vocab}
\title{The Birds of Prey dataset}
\format{\code{bop.vocab} a vector of unique words in the corpus vocabulary.

\code{bop.docs} a list of documents in the corpus. Each item (represents a
 document) is a matrix (2 X U) of word frequencies, where U represents the
 number of unique words in a document. Each column in the matrix represents
 a unique word in a document and contains
 \itemize{
   \item vocabulary-id. the index of the word in the vocabulary (starts with 0)
   \item frequency. the relative frequency of the word in the document
 }

\code{bop.docs.metadata} a matrix of document (article) metadata, where
each row represents a document with
 \itemize{
   \item category. the Wikipedia category assigned to the article
   \item title. the title of the Wikipedia web article
 }}
\source{
The Wikipedia articles are downloaded from the
 \href{http://en.wikipedia.org/wiki/Main_Page}{English Wikipedia}
 with the help of
 \href{http://www.mediawiki.org/wiki/API:Main_page}{Media Wiki} API.
}
\usage{
\code{data(bop.vocab)}
\code{data(bop.docs)}
\code{data(bop.docs.metadata)}
}
\description{
A subset of Wikipedia artricles under the Wikipedia category
\href{http://en.wikipedia.org/wiki/Category:Birds_of_prey}{Birds
of Prey} formatted for running the LDA Gibbs sampling
algorithms. This dataset contains 304 Wikipedia articles from
the following Wikipedia categories.
 \itemize{
   \item Category:Eagles
   \item Category:Falco (genus)
   \item Category:Falconry
   \item Category:Falcons
   \item Category:Harriers (birds)
   \item Category:Hawks
   \item Category:Kites (birds)
   \item Category:Owls
 }
}
\author{
Clint P. George, April 15, 2014
}
\seealso{
Other datasets: \code{\link{whales}},
  \code{\link{whales.docs}},
  \code{\link{whales.docs.metadata}},
  \code{\link{whales.vocab}}; \code{\link{wt16}},
  \code{\link{wt16.docs}},
  \code{\link{wt16.docs.metadata}},
  \code{\link{wt16.vocab}}; \code{\link{wt}},
  \code{\link{wt.docs}}, \code{\link{wt.docs.metadata}},
  \code{\link{wt.vocab}}
}

