% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_beta_topic_labels.R
\name{calc_beta_topic_labels}
\alias{calc_beta_topic_labels}
\title{Computes \eqn{\beta} topic labels}
\usage{
calc_beta_topic_labels(true.beta, mc.betas)
}
\arguments{
\item{true.beta}{the \code{true} \eqn{\beta} (a \eqn{K \times V} matrix)
computed from the corpus using \code{true} document topic labels and 
\code{\link{calc_class_term_frequency}}, where \eqn{K} represents the 
number of topics in the corpus and \eqn{V} represents the number of terms 
in the vocabulary.}

\item{mc.betas}{a \eqn{K \times V \times S} matrix, which has the 
\eqn{\beta^(s), s = 1, \ldots, S} samples from a FGS or ACGS chain}
}
\value{
A \eqn{K \times S} matrix of \eqn{\beta_j^(s)} topic labels
}
\description{
Computes topic lables for each \eqn{\beta_j^(s), s = 1, \ldots, S} sampled 
from the FGS or ACGS chain
}
\seealso{
\code{\link{calc_class_term_frequency}}, \code{\link{normalize}}
}

