% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc_class_term_frequency}
\alias{calc_class_term_frequency}
\title{Calculate class term frequency matrix}
\usage{
calc_class_term_frequency(class.labels, vocab, doc.class.labels, documents)
}
\arguments{
\item{class.labels}{A vector of true class names}

\item{vocab}{a vector of words in the vocabulary}

\item{doc.class.labels}{a vector of document classes}

\item{documents}{a list documents read via \code{\link{read_docs}}}
}
\value{
The class term frequency matrix
}
\description{
Computes the class term frequency matrix for a given corpus.
}
\examples{
documents <- read_docs('bop.ldac');
vocab <- readLines('bop.ldac.vocab');

doc.metadata <- read.csv2('bop.csv', header = T, sep = ';');
class.labels <- levels(doc.metadata[, 'category'])
ctf <- calc_class_term_frequency(class.labels, vocab,
                                 doc.metadata[, 'category'],
                                 documents);
}

