% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_doc_cos.R
\name{calc_doc_cos}
\alias{calc_doc_cos}
\title{Computes cosine scores between documents in a corpus}
\usage{
calc_doc_cos(docs)
}
\arguments{
\item{docs}{corpus document matrix, a \eqn{D \times M} matrix where D is the 
number of documents in the corpus and M is the number of features}
}
\value{
sorted document index pairs in terms of pairwise cosine scores 
(\eqn{4 \times (D*D/2)} matrix)
}
\description{
Computes cosine scores between documents in a corpus
}
\details{
Last modified on: May 24, 2015
}
\seealso{
Other corpus: \code{\link{calc_doc_tf}},
  \code{\link{gen_corpus}}
}
