% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ldac_documents.R
\name{calc_doc_lengths}
\alias{calc_doc_lengths}
\title{Calculate document sizes}
\usage{
calc_doc_lengths(docs)
}
\arguments{
\item{docs}{a list of documents}
}
\value{
a list of document lengths
}
\description{
Computes the number of words in each document that are read 
via \code{\link{read_docs}}.
}
\examples{
documents <- read_docs('bop.ldac');
ds <- vectorize_docs(documents);
doc.N <- calc_doc_lengths(documents)

}
\seealso{
Other lda data preprocessing methods: \code{\link{read_docs}},
  \code{\link{vectorize_docs}}
}

