% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_doc_tf.R
\name{calc_doc_tf}
\alias{calc_doc_tf}
\title{Computes term frequency (tf) vectors for documents}
\usage{
calc_doc_tf(D, V, word.ids, doc.ids)
}
\arguments{
\item{D}{number of documents}

\item{V}{vocabulary size}

\item{wid}{word ids of document words, generated via \code{gen_corpus}}

\item{did}{document ids of document words, generated via \code{gen_corpus}}
}
\value{
tf matrix (\eqn{D \times V})
}
\description{
Computes term frequency (tf) vectors for documents
}
\details{
Last modified on: May 24, 2015
}
\seealso{
Other corpus: \code{\link{calc_doc_cos}}
}

