% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.R
\name{calc_tfidf}
\alias{calc_tfidf}
\title{Calculate TF-IDF values}
\usage{
calc_tfidf(documents, vocab)
}
\arguments{
\item{documents}{a list documents read using \code{\link{read_docs}}}

\item{vocab}{a list of unique words in the vocabulary}
}
\value{
A list of TF-IDF vectors for each document in the corpus
}
\description{
Computes the TF-IDF model for a given corpus.
}
\examples{
documents <- read_docs('bop.ldac');
vocab <- readLines(vocab.file);
tfidfs <- calc_tfidf(documents, vocab);


}

