% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_top_topic_words.R
\name{calc_top_topic_words}
\alias{calc_top_topic_words}
\title{Gets the most probable topical words}
\usage{
calc_top_topic_words(beta, vocab, num.words = 30, num.digits = 2)
}
\arguments{
\item{beta}{the \eqn{\beta} matrix in the LDA model, which is obtained from any LDA Gibbs sampler}

\item{vocab}{the terms in the corpus vocabulary as a list. This should follow the same order of beta}

\item{num.words}{the number of most probabale words to display. The default is 30 words.}

\item{num.digits}{the number of decimal digits to be displayed for the probabilities}
}
\description{
Returns \code{num.words}-most probable words for each topic in the LDA 
model learned for a corpus
}
\examples{

calc_top_topic_words(beta, vocab, num.words=30, num.digits=2)
              
 
}
\seealso{
\code{\link{lda_fgs}}, \code{\link{lda_acgs}}, \code{\link{lda_fgs_blei_corpus}}
}

