% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cats}
\alias{cats}
\alias{cats.class.labels}
\alias{cats.doc.N}
\alias{cats.docs}
\alias{cats.docs.metadata}
\alias{cats.ds}
\alias{cats.vocab}
\title{Cats (C-6)}
\format{\code{vocab} a vector of unique words in the corpus vocabulary.
 
\code{docs} a list of documents in the corpus. Each item (represents a 
 document) is a matrix (2 X U) of word frequencies, where U represents the 
 number of unique words in a document. Each column in the matrix represents 
 a unique word in a document and contains    
 \itemize{
   \item vocabulary-id. the index of the word in the vocabulary (starts with 0)  
   \item frequency. the relative frequency of the word in the document   
 }         
 
\code{docs.metadata} a matrix of document (article) metadata, where each 
row represents a document with        
 \itemize{
   \item category. the Wikipedia category assigned to the article 
   \item title. the title of the Wikipedia web article     
 } 
 
\code{doc.N} a vector of word counts of documents in the corpus 

\code{num.docs} the number of documents in the corpus 

\code{class.labels} a vector of unique categories (classes) in the corpus 

\code{ds.name} the corpus name (string)

\code{ds} a list of two equal-length vectors       
 \itemize{
   \item wid. vocabulary ids of the instances of words in the corpus (a 
   vector)
   \item did. document indices of the instances of words in the corpus (a 
   vector)  
 }}
\source{
Articles are downloaded from the 
 \href{http://en.wikipedia.org/wiki/Main_Page}{English Wikipedia}
 with the help of 
 \href{http://www.mediawiki.org/wiki/API:Main_page}{Media Wiki} API.
}
\usage{
\code{data(cats)}
}
\description{
A corpus created from a subset of the Wikipedia articles under the 
categories: 
 \itemize{
   \item Leopardus (8 documents)
   \item Lynx (8 documents)
   \item Prionailurus (7 documents).
 }
All the three categories of this corpus are under the Wikipedia 
super-category \href{http://en.wikipedia.org/wiki/Category:Felines}{Felines}.
}
\note{
Created on November 21, 2015
}
\author{
Clint P. George
}
\seealso{
Other datasets: \code{\link{autos-motorcycles}},
  \code{\link{bop}}, \code{\link{canis}},
  \code{\link{felines}}, \code{\link{ibm-mac}},
  \code{\link{med-christian-baseball}}, \code{\link{rec}},
  \code{\link{sci}}, \code{\link{whales}},
  \code{\link{wt16m}}, \code{\link{wt16}}, \code{\link{wt}}
}

