% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_synth_corpus.R
\name{gen_corpus}
\alias{gen_corpus}
\title{Generates a synthetic corpus based on symmetric Dirichlets}
\usage{
gen_corpus(K, V, D, doc.size, alpha, eta)
}
\arguments{
\item{K}{number of topics}

\item{V}{vocabulary size}

\item{D}{number of documents}

\item{doc.size}{number of words in each document}

\item{alpha}{hyperparameter for document Dirichlet sampling}

\item{eta}{hyperparameter for topic Diriclet sampling}
}
\value{
a list of generated documents and their statistics
}
\description{
Generates documents using the LDA generative process based on a set of 
predefined values.
}
\details{
Last modified on: May 24, 2015
}
\examples{
## Generates documents with given parameters 

K              <- 2
V              <- 20 
D              <- 100
alpha          <- 7
eta            <- 3
doc.size       <- 80

ds             <- gen_corpus(K, V, D, doc.size, alpha, eta);

}
\seealso{
Other corpus: \code{\link{calc_doc_cos}},
  \code{\link{calc_doc_tf}}
}

