% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_synth_corpus.R
\name{gen_synth_corpus}
\alias{gen_synth_corpus}
\title{Generates a synthetic corpus}
\usage{
gen_synth_corpus(D, lambda.hat, alpha.v, beta)
}
\arguments{
\item{D}{the number of documents in the corpus}

\item{lambda.hat}{the mean of document counts}

\item{alpha.v}{the vector of Dirichlet hyperparameters (K X 1) for document
topic mixtures}

\item{beta}{the beta matrix (counts) for topic word probabilities (K x V
format)}
}
\value{
a list of generated documents' details
}
\description{
Generates document words using the LDA generative process given a beta. It's
used to test the correctness of the Gibbs sampling algorithms.
}
\examples{
K              <- 2
V              <- 20 
D              <- 100
gen.alpha.v    <- array(7, c(K, 1)); 
gen.eta.v      <- array(3, c(1, V)); 
lambda.hat     <- 80

## Generates the synthetic beta.m
beta.m         <- matrix(1e-2, nrow=K, ncol=V)
beta.m[1, ]    <- rdirichlet(1, gen.eta.v);
beta.m[2, ]    <- rdirichlet(1, gen.eta.v);

## Generates documents with a given beta.m
ds             <- gen_synth_corpus(D, lambda.hat, gen.alpha.v, beta.m);

}
