% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lda_acgs}
\alias{lda_acgs}
\title{The LDA Augmented Collapsed Gibbs sampler (ACGS)}
\usage{
lda_acgs(K, V, wid, doc.N, alpha.v, eta, max.iter = 100, burn.in = 0,
  spacing = 1, store.Dir = 1)
}
\arguments{
\item{K}{the number of topics in the corpus}

\item{V}{the vocabulary size}

\item{wid}{the vocabulary ids of every word instance in each corpus document
(1 X total.N vector). We assume vocabulary id starts with 1}

\item{doc.N}{the document lengths}

\item{alpha.v}{the hyper parameter vector for \eqn{\theta}}

\item{eta}{the \eqn{\beta} matrix smoothing parameter}

\item{max.iter}{the max number of Gibbs iterations to be performed}

\item{burn.in}{the burn in period of the Gibbs sampler}

\item{spacing}{the spacing between the stored samples (to reduce correlation)}

\item{store.Dir}{if 0 the sampler does not save \eqn{(\theta, \beta)} samples}
}
\value{
the Gibbs sampling output
}
\description{
This a Markov chain on \eqn{(z, \beta, \theta)} extending the Collapsed Gibbs
Sampler of Griffiths and Steyvers (2004), which is a Markov chain on \eqn{z}
}
\seealso{
Other Gibbs.sampling.methods: \code{\link{lda_fgs_blei_corpus}};
  \code{\link{lda_fgs}}
}

