% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_acgs_hs.R
\name{lda_acgs_hs}
\alias{lda_acgs_hs}
\title{The LDA Augmented Collapsed Gibbs sampler (ACGS) with Selection of h}
\usage{
lda_acgs_hs(K, V, wid, doc.N, alpha, eta, h.grid, max.iter, burn.in, spacing,
  save.z, save.beta, save.theta, save.Bh, save.lp)
}
\arguments{
\item{K}{the number of topics in the corpus}

\item{V}{the vocabulary size}

\item{wid}{the vocabulary ids of every word instance in each corpus document
(1 X total.N vector). We assume vocabulary id starts with 1}

\item{doc.N}{the document lengths}

\item{alpha}{the hyper parameter for \eqn{\theta}}

\item{eta}{the \eqn{\beta} matrix smoothing parameter}

\item{h.grid}{the grid of hyperparamets \eqn{(\eta, \alpha)}, 2 x G matrix,
where G is the number of grid points and the first row is for \eqn{\alpha}
points and the second row is for \eqn{\eta} points}

\item{max.iter}{the max number of Gibbs iterations to be performed}

\item{burn.in}{the burn in period of the Gibbs sampler}

\item{spacing}{the spacing between the stored samples (to reduce correlation)}

\item{save.z}{if 0 the function does not save \eqn{z} samples}

\item{save.beta}{if 0 the function does not save \eqn{\beta} samples}

\item{save.theta}{if 0 the function does not save \eqn{\theta} samples}

\item{save.Bh}{if 0 the function does not save computed \eqn{B(h, h_*)}
values for iterations}

\item{save.lp}{if 0 the function does not save computed log posterior for
iterations}
}
\value{
the Gibbs sampling output
}
\description{
This a Markov chain on \eqn{(z, \beta, \theta)} extending the Collapsed Gibbs
Sampler of Griffiths and Steyvers (2004)
}
\seealso{
Other Gibbs sampling methods: \code{\link{get_grid_neighbors}},
  \code{\link{lda_acgs}},
  \code{\link{lda_fgs_blei_corpus}},
  \code{\link{lda_fgs_hs}}, \code{\link{lda_fgs_st}},
  \code{\link{lda_fgs}}
}

