% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lda_cgs_em}
\alias{lda_cgs_em}
\title{LDA: Gibbs-EM}
\usage{
lda_cgs_em(num_topics, vocab_size, docs_tf, alpha_h, eta_h, em_max_iter,
  gibbs_max_iter, burn_in, spacing, verbose)
}
\description{
This implements the Gibbs-EM algorithm for LDA that is mentioned in the  
paper Topic Modeling: Beyond Bag-of-Words. Wallach (2006).
}
\seealso{
Other MCMC: \code{\link{lda_acgs_st}},
  \code{\link{lda_cgs_em_perplexity}},
  \code{\link{lda_cgs_perplexity}},
  \code{\link{lda_fgs_BF_perplexity}},
  \code{\link{lda_fgs_perplexity}},
  \code{\link{lda_fgs_ppc}},
  \code{\link{lda_fgs_st_perplexity}}
}
