% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lda_fgs}
\alias{lda_fgs}
\title{Full Gibbs sampler (FGS) for LDA}
\usage{
lda_fgs(K, V, wid, doc.N, alpha.v, eta, max.iter = 100, burn.in = 0,
  spacing = 1, store.Dir = 1)
}
\arguments{
\item{K}{Number of topics in the corpus}

\item{V}{Vocabulary size}

\item{wid}{Vocabulary ids of every word instance in each corpus document
(1 X total.N vector). We assume vocabulary id starts with 1}

\item{doc.N}{Documents' word counts}

\item{alpha.v}{Hyperparameter vector for \eqn{\theta}}

\item{eta}{Smoothing parameter for the \eqn{\beta} matrix}

\item{max.iter}{Maximum number of Gibbs iterations to be performed}

\item{burn.in}{Burn-in-period for the Gibbs sampler}

\item{spacing}{Spacing between the stored samples (to reduce correlation)}

\item{store.Dir}{If 0 the sampler does not save \eqn{(\theta, \beta)} samples}
}
\value{
the Gibbs sampling output
}
\description{
This a R wrapper function for the C++ implementation of the full Gibbs
sampler for LDA---a Markov chain on \eqn{(z, \beta, \theta)}.
}
\seealso{
Other Gibbs.sampling.methods: \code{\link{lda_acgs}};
  \code{\link{lda_fgs_blei_corpus}}
}

