% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lda_fgs_blei_corpus}
\alias{lda_fgs_blei_corpus}
\title{The LDA Full Gibbs sampler (FGS)}
\usage{
lda_fgs_blei_corpus(K, V, doc.N, docs, alpha.v, eta, max.iter = 100,
  burn.in = 0, spacing = 1, store.Dir = 1)
}
\arguments{
\item{K}{the number of topics in the corpus}

\item{V}{the vocabulary size}

\item{doc.N}{the document lengths}

\item{docs}{a list of corpus documents read from the Blei corpus}

\item{alpha.v}{the hyper parameter vector for \eqn{\theta}}

\item{eta}{the \eqn{\beta} matrix smoothing parameter}

\item{max.iter}{the max number of Gibbs iterations to be performed}

\item{burn.in}{the burn in period of the Gibbs sampler}

\item{spacing}{the spacing between the stored samples (to reduce correlation)}

\item{store.Dir}{if 0 the sampler does not save \eqn{(\theta, \beta)} samples}
}
\value{
the Gibbs sampling output
}
\description{
This a R wrapper function for the C++ implementation of the full Gibbs
sampler for LDA, which is a Markov chain on \eqn{(z, \beta, \theta)}. This
function uses a corpus that is created using \code{\link{read_docs}}.
}
\seealso{
Other Gibbs.sampling.methods: \code{\link{lda_acgs}};
  \code{\link{lda_fgs}}
}

