% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lda_fgs_lppv.R
\name{lda_fgs_lppv}
\alias{lda_fgs_lppv}
\title{Computes Log Posterior Predictive Value for an LDA model:}
\usage{
lda_fgs_lppv(K, V, wid, doc.N, alpha.v, eta, max.iter = 100, burn.in = 0,
  spacing = 1)
}
\arguments{
\item{K}{Number of topics in the corpus}

\item{V}{Vocabulary size}

\item{wid}{a vector of vocabulary ids of every word instance in the corpus.
Note: we assume vocabulary id starts with 1.}

\item{doc.N}{a vector of word counts for each document in the corpus}

\item{alpha.v}{hyperparameter vector for document Dirichlets \eqn{\theta}}

\item{eta}{hyperparameter value for topic Dirichlets \eqn{\beta}}

\item{max.iter}{maximum number of Gibbs iterations to be performed}

\item{burn.in}{burn-in period for the Gibbs sampler}

\item{spacing}{spacing between the stored samples (to reduce correlation)}
}
\value{
A list that consists of
  (a) a (D x S) matrix of log posterior predictive values for each held-out
      using each sample, where S is the number of saved samples based on
      burn.in and spacing, and
  (b) log posterior predictive value of the corpus.
}
\description{
This is based on Zhe Chen (2015) [dissertation].
}
\seealso{
Other posterior predictive check (PPC) options: \code{\link{lda_acgs_lppv_R}};
  \code{\link{lda_fgs_lppv_R}}
}

