% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_fgs_lppv_R.R
\name{lda_fgs_lppv_R}
\alias{lda_fgs_lppv_R}
\title{Log Posterior Predictive Value based on the Full Gibbs Sampler}
\usage{
lda_fgs_lppv_R(K, V, alpha, eta, did, wid, doc.N, max.iter = 5 * 10^3,
  burn.in = 10^3, spacing = 1)
}
\arguments{
\item{K}{Number of topics in the corpus}

\item{V}{Vocabulary size}

\item{alpha}{Hyperparameter value for \eqn{\theta} matrix}

\item{eta}{Smoothing parameter for the \eqn{\beta} matrix}

\item{did}{Document ids of every word instance in each corpus document
(1 X total.N vector). We assume vocabulary id starts with 1}

\item{wid}{Vocabulary ids of every word instance in each corpus document
(1 X total.N vector). We assume vocabulary id starts with 1}

\item{doc.N}{Documents' word counts}

\item{max.iter}{Maximum number of Gibbs iterations to be performed}

\item{burn.in}{Burn-in-period for the Gibbs sampler}

\item{spacing}{Spacing between the stored samples (to reduce correlation)}
}
\value{
Log Posterior Predictive Value
}
\description{
Computation is based on Zhe Chen (2015)'s method
}
\seealso{
Other posterior predictive check (PPC) options: \code{\link{lda_acgs_lppv_R}},
  \code{\link{lda_fgs_lppv}}
}

