% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_fgs_st.R
\name{lda_fgs_st}
\alias{lda_fgs_st}
\title{LDA Full Gibbs sampler (FGS) with Serial Tempering and Tuning}
\usage{
lda_fgs_st(K, V, wid, doc.N, h.grid, st.grid, st.grid.nbrs, init.st.grid.index,
  init.st.grid.zetas, max.iter, burn.in, spacing, tuning.iter, save.z,
  save.beta, save.theta, save.st.grid.index, save.lp, save.hat.ratios,
  save.tilde.ratios, verbose, max.iter.final = max.iter)
}
\arguments{
\item{K}{the number of topics in the corpus}

\item{V}{the vocabulary size}

\item{wid}{the vocabulary ids of every word instance in each corpus document 
(1 X total.N vector). We assume vocabulary id starts with 1}

\item{doc.N}{the document lengths}

\item{h.grid}{the grid of hyperparameters \eqn{h = (\eta, \alpha)}, 2 x G
matrix, where G is the number of grid points and the first row is for
\eqn{\alpha} points and the second row is for \eqn{\eta} points}

\item{st.grid}{the helper grid of hyperparameters \eqn{h = (\eta, \alpha)}, 
2 x G' matrix, for Serial Tempering, where G' is the number of grid points 
and the first row is for \eqn{\alpha} points and the second row is for 
\eqn{\eta} points}

\item{st.grid.nbrs}{the neighbor indices, from \eqn{[0, G']}, of each
st.grid point}

\item{init.st.grid.index}{the initial h index from st.grid for Serial Tempering}

\item{init.st.grid.zetas}{initial guess for normalization constants}

\item{max.iter}{the max number of Gibbs iterations to be performed}

\item{burn.in}{the burn in period of the Gibbs sampler}

\item{spacing}{the spacing between the stored samples (to reduce correlation)}

\item{tuning.iter}{the number of tuning iterations}

\item{save.z}{if 0 the function does not save \eqn{z} samples}

\item{save.beta}{if 0 the function does not save \eqn{\beta} samples}

\item{save.theta}{if 0 the function does not save \eqn{\theta} samples}

\item{save.st.grid.index}{if 0 the function does not save selected indices 
of h in the grid}

\item{save.lp}{if 0 the function does not save computed log posterior for
iterations}

\item{save.hat.ratios}{if 1 sub grid and main grid ratios are saved}

\item{save.tilde.ratios}{if 1 normalized main grid ratios are saved. It's 
mainly used for estimating posterior expectations.}

\item{verbose}{values: {1, 0}}

\item{max.iter.final}{the max number of Gibbs iterations to be performed for 
the final tuning run}
}
\value{
the Gibbs sampling output
}
\description{
This a Markov chain on \eqn{(z, \beta, \theta)}
}
\details{
Last modified on: January 28, 2016
}
\seealso{
Other Gibbs sampling methods: \code{\link{get_grid_neighbors}},
  \code{\link{lda_acgs_hs}}, \code{\link{lda_acgs}},
  \code{\link{lda_fgs_hs}}, \code{\link{lda_fgs}}
}
