% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ldamcmc}
\alias{ldamcmc}
\alias{ldamcmc-package}
\alias{package-ldamcmc}
\title{Implementations of Gibbs sampling algorithms for the Latent Dirichlet
Allocation Model}
\description{
TODO
}
\references{
Clint P. George and Hani Doss (2014). Principled Selections of
Hyperparameter in the Latent Dirichlet Allocation Model.
}

