% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{Normalizes a given matrix}
\usage{
normalize(XX, dim = 1)
}
\arguments{
\item{XX}{a 2-dimensional matrix to normalize.}

\item{dim}{the normalizing dimension (\code{1} column wise, \code{2} row wise).}
}
\value{
The normalized input matrix.
}
\description{
Normalizes a given matrix
}
\examples{
XX <- array(2, c(3, 2));
normalize(XX, dim=2);
}
\seealso{
\code{\link{colSums}} 
\code{\link{rowSums}}
}

