% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_meshgrid}
\alias{plot_meshgrid}
\title{Plot a meshgrid}
\usage{
plot_meshgrid(values, x.axis, y.axis, xlabel, ylabel, zlabel, main.title = "",
  plot.file = "", surface.color = "lightblue", zlim = range(values, na.rm
  = TRUE))
}
\arguments{
\item{values}{the z values}

\item{x.axis,y.axis}{the \code{x-y} coordinates at which z is evaluated}

\item{xlabel,ylabel,zlabel}{the axis labels of \code{x, y, z}}

\item{main.title}{the main title of the plot (optional)}

\item{plot.file}{the file path at which the plot to be saved (optional)}

\item{surface.color}{the plot surface color. The default is light blue.}

\item{zlim}{the range of \code{z} values for the \code{\link{persp}} plot.
  The default is \code{range(values, na.rm=TRUE)}}
}
\description{
Plots a mesh over the \code{x-y} grid with the help of
\code{\link[graphics]{persp}} function
}
\examples{
data(meshgrid);

plot_meshgrid(values, x.axis, y.axis, "alpha", "eta", "Estimate of B(h)");
}
\seealso{
\code{\link{persp}} \code{\link[lattice]{trellis.device}}
}

