% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ldac_documents.R
\name{read_docs}
\alias{read_docs}
\title{Read documents}
\usage{
read_docs(filename)
}
\arguments{
\item{filename}{the LDA-C formatted file}
}
\value{
A list of documents (the vocab id starts at 0)
}
\description{
A function to read documents from a given LDA-C formatted file.
}
\examples{
documents <- read_docs('bop.ldac');

}
\references{
The \pkg{\link{lda}} R package by Chong Wong
}
\seealso{
\pkg{\link{lda}}

Other lda data preprocessing methods: \code{\link{calc_doc_lengths}},
  \code{\link{vectorize_docs}}
}

