% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_antoniak}
\alias{sample_antoniak}
\title{Samples from the Antoniak distribution}
\usage{
sample_antoniak(N, alpha)
}
\arguments{
\item{N}{Number of samples}

\item{alpha}{strength parameter}
}
\description{
It's done by sampling \eqn{N} Bernoulli variables
}
\details{
References:

http://www.jmlr.org/papers/volume10/newman09a/newman09a.pdf
}
\note{
Created on: May 19, 2016

Created by: Clint P. George
}
\seealso{
Other utils: \code{\link{sample_dirichlet}},
  \code{\link{sample_multinomial}}
}
