% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_dirichlet}
\alias{sample_dirichlet}
\title{Samples from a Dirichlet distribution given a hyperparameter}
\usage{
sample_dirichlet(num_elements, alpha)
}
\arguments{
\item{num_elements}{the dimention of the Dirichlet distribution}

\item{alpha}{the hyperparameter vector (a column vector)}
}
\value{
returns a Dirichlet sample (a column vector)
}
\description{
Samples from a Dirichlet distribution given a hyperparameter
}
\note{
Author: Clint P. George

Created on: 2014
}
\seealso{
Other utils: \code{\link{sample_antoniak}},
  \code{\link{sample_multinomial}}
}
