% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_multinomial}
\alias{sample_multinomial}
\title{A speedy sampling from a multimomial distribution}
\usage{
sample_multinomial(theta)
}
\arguments{
\item{theta}{a multinomial probability vector (K x 1 vector)}
}
\value{
returns a class index from [0, K)
}
\description{
A speedy sampling from a multimomial distribution
}
\note{
Author: Clint P. George

Created on: February 11, 2016
}
\seealso{
Other utils: \code{\link{sample_antoniak}},
  \code{\link{sample_dirichlet}}
}
