% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sci}
\alias{sci}
\alias{sci.class.labels}
\alias{sci.doc.N}
\alias{sci.docs}
\alias{sci.docs.metadata}
\alias{sci.ds}
\alias{sci.vocab}
\title{Science (C-3)}
\format{\code{vocab} a vector of unique words in the corpus vocabulary.
 
\code{docs} a list of documents in the corpus. Each item (represents a 
 document) is a matrix (2 X U) of word frequencies, where U represents the 
 number of unique words in a document. Each column in the matrix represents 
 a unique word in a document and contains    
 \itemize{
   \item vocabulary-id. the index of the word in the vocabulary (starts with 0)  
   \item frequency. the relative frequency of the word in the document   
 }         
 
\code{docs.metadata} a matrix of document (article) metadata, where each 
row represents a document with        
 \itemize{
   \item category. the Wikipedia category assigned to the article 
   \item title. the title of the Wikipedia web article     
 } 
 
\code{doc.N} a vector of word counts of documents in the corpus 

\code{num.docs} the number of documents in the corpus 

\code{class.labels} a vector of unique categories (classes) in the corpus 

\code{ds.name} the corpus name (string)

\code{ds} a list of two equal-length vectors       
 \itemize{
   \item wid. vocabulary ids of the instances of words in the corpus (a 
   vector)
   \item did. document indices of the instances of words in the corpus (a 
   vector)  
 }}
\source{
Articles and categories are adapted from the     
 \href{http://qwone.com/~jason/20Newsgroups}{20Newsgroups} dataset.
}
\usage{
\code{data("sci")}
}
\description{
A corpus created from the 20Newsgroups dataset. This corpus is created from 
a random subset of articles from the 20Newsgroups categories:
 \itemize{
   \item Cryptography (50 documents)
   \item Electronics (50 documents) 
   \item Medicine (50 documents)
   \item Space (50 documents).
 }
All four of these categories are classified 
under the super-category Science in the 20Newsgroups dataset.
}
\note{
Created on November 21, 2015
}
\author{
Clint P. George
}
\seealso{
Other datasets: \code{\link{autos-motorcycles}},
  \code{\link{bop}}, \code{\link{canis}},
  \code{\link{cats}}, \code{\link{felines}},
  \code{\link{ibm-mac}},
  \code{\link{med-christian-baseball}}, \code{\link{rec}},
  \code{\link{whales}}, \code{\link{wt16m}},
  \code{\link{wt16}}, \code{\link{wt}}
}

