% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ldac_documents.R
\name{vectorize_docs}
\alias{vectorize_docs}
\title{Vectorize the documents}
\usage{
vectorize_docs(docs)
}
\arguments{
\item{docs}{a list of documents, which is created using
\code{\link{read_docs}}}
}
\value{
A list of document and word instances
}
\description{
Converts the documents read using \code{\link{read_docs}} into two vectors:
one vector for the document word instances (contains vocabulary id's) and the
other vector for the corresponding document id's.
}
\note{
This method is very time consuming for large datasets. Therefore, use 
  functions such as \code{\link{lda_fgs_blei_corpus}}, which take \code{docs}
  as input and do the job of this function in the C++ programming langauge, 
  for Gibbs sampling.
}
\examples{
documents <- read_docs('bop.ldac');
ds <- vectorize_docs(documents);

}
\seealso{
\code{\link{lda_fgs_blei_corpus}}

Other lda data preprocessing methods: \code{\link{calc_doc_lengths}},
  \code{\link{read_docs}}
}

