% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{whales}
\alias{whales}
\alias{whales.docs}
\alias{whales.docs.metadata}
\alias{whales.vocab}
\title{The Whales dataset}
\format{\code{whales.vocab} a vector of unique words in the corpus vocabulary.
 
\code{whales.docs} a list of documents in the corpus. Each item (represents a 
 document) is a matrix (2 X U) of word frequencies, where U represents the 
 number of unique words in a document. Each column in the matrix represents 
 a unique word in a document and contains    
 \itemize{
   \item vocabulary-id. the index of the word in the vocabulary (starts with 0)  
   \item frequency. the relative frequency of the word in the document   
 }          
 
\code{whales.docs.metadata} a matrix of document (article) metadata, where 
each row represents a document with        
 \itemize{
   \item category. the Wikipedia category assigned to the article 
   \item title. the title of the Wikipedia web article     
 }}
\source{
The Wikipedia articles are downloaded from the 
 \href{http://en.wikipedia.org/wiki/Main_Page}{English Wikipedia}
 with the help of 
 \href{http://www.mediawiki.org/wiki/API:Main_page}{Media Wiki} API.
}
\usage{
\code{data(whales.vocab)}
\code{data(whales.docs)}
\code{data(whales.docs.metadata)}
}
\description{
A subset of Wikipedia artricles under the Wikipedia category 
\href{http://en.wikipedia.org/wiki/Category:Whales}{Whales} 
formatted for running the LDA Gibbs sampling algorithms. This 
dataset contains 153 Wikipedia articles from the following 
Wikipedia subcategories.
 \itemize{
   \item Category:Baleen whales 
   \item Category:Dolphins 
   \item Category:Killer whales 
   \item Category:Oceanic dolphins 
   \item Category:Whale products          
   \item Category:Whaling
 }
}
\author{
Clint P. George, April 15, 2014
}
\seealso{
Other datasets: \code{\link{bop}}, \code{\link{wt16}},
  \code{\link{wt}}
}

