% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{wt}
\alias{wt}
\alias{wt.docs}
\alias{wt.docs.metadata}
\alias{wt.vocab}
\title{The Whales and Tires dataset}
\format{\code{wt.vocab} a vector of unique words in the corpus vocabulary.

\code{wt.docs} a list of documents in the corpus, where each
 document is represented by a matrix (2 X U) of word frequencies.
 The variable U is the number of unique words in a document. Each
 matrix column represents a unique word in a document which
 contains the following variables
 \itemize{
   \item vocabulary-id. the index of a term in the vocabulary (
   starts with 0)
   \item frequency. the relative of a term in a given document
 }

\code{wt.docs.metadata} a matrix of documents metadata, where
each row that represents a document contains
 \itemize{
   \item category. the Wikipedia category assigned to an article
   \item title. the title of the Wikipedia web article
 }}
\source{
The articles are downloaded from the
 \href{http://en.wikipedia.org/wiki/Main_Page}{English Wikipedia}
 with the help of
 \href{http://www.mediawiki.org/wiki/API:Main_page}{Media Wiki}
}
\usage{
\code{data(wt.vocab)}
\code{data(wt.docs)}
\code{data(wt.docs.metadata)}
}
\description{
A subset of Wikipedia artricles under the Wikipedia categories
\href{http://en.wikipedia.org/wiki/Category:Whales}{Whales}
and \href{http://en.wikipedia.org/wiki/Category:Tires}{Tires}
formatted for running the LDA Gibbs sampling algorithms. This
dataset contains 84 Wikipedia articles.
}
\author{
Clint P. George, April 15, 2014
}
\seealso{
\pkg{\link{lda}}

Other datasets: \code{\link{bop}}, \code{\link{bop.docs}},
  \code{\link{bop.docs.metadata}}, \code{\link{bop.vocab}};
  \code{\link{whales}}, \code{\link{whales.docs}},
  \code{\link{whales.docs.metadata}},
  \code{\link{whales.vocab}}
}

