% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lda_acgs.R
\name{lda_acgs}
\alias{lda_acgs}
\title{Augmented collapsed Gibbs sampler (ACGS) for LDA}
\usage{
lda_acgs(K, V, wid, doc.N, alpha.v, eta, max.iter = 100, burn.in = 0,
  spacing = 1, save.z = 0, save.beta = 0, save.theta = 0, save.lp = 0)
}
\arguments{
\item{K}{Number of topics in the corpus}

\item{V}{Vocabulary size}

\item{wid}{Vocabulary ids of every word instance in each corpus document
(1 X total.N vector). We assume vocabulary id starts with 1}

\item{doc.N}{Documents' word counts}

\item{alpha.v}{Hyperparameter vector for \eqn{\theta}}

\item{eta}{Smoothing parameter for the \eqn{\beta} matrix}

\item{max.iter}{Maximum number of Gibbs iterations to be performed}

\item{burn.in}{Burn-in-period for the Gibbs sampler}

\item{spacing}{Spacing between the stored samples (to reduce correlation)}

\item{save.z}{if 0 the function does not save \eqn{z} samples}

\item{save.beta}{if 0 the function does not save \eqn{\beta} samples}

\item{save.theta}{if 0 the function does not save \eqn{\theta} samples}

\item{save.lp}{if 0 the function does not save computed log posterior for
           iterations}
}
\value{
the Gibbs sampling output
}
\description{
This a Markov chain on \eqn{(z, \beta, \theta)} extending the collapsed Gibbs
sampler (CGS) of Griffiths and Steyvers (2004)---a Markov chain on \eqn{z}.
}
\seealso{
Other Gibbs.sampling.methods: \code{\link{lda_fgs_blei_corpus}};
  \code{\link{lda_fgs}}
}

