% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ldamcmc.R
\docType{package}
\name{ldamcmc}
\alias{ldamcmc}
\alias{ldamcmc-package}
\alias{package-ldamcmc}
\title{Implementations of Gibbs sampling algorithms for the Latent Dirichlet
Allocation Model}
\description{
TODO
}
\references{
Clint P. George and Hani Doss (2014). Principled Selection of Hyperparameters
in the Latent Dirichlet Allocation Model.
}

