% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{whales}
\alias{whales}
\alias{whales.docs}
\alias{whales.docs.metadata}
\alias{whales.vocab}
\title{The Whales dataset}
\format{\code{whales.vocab} a vector of unique words in the corpus vocabulary.

\code{whales.docs} a list of documents in the corpus, where each
 document is represented by a matrix (2 X U) of word frequencies.
 The variable U is the number of unique words in a document. Each
 matrix column represents a unique word in a document which
 contains the following variables
 \itemize{
   \item vocabulary-id. the index of a term in the vocabulary (
   starts with 0)
   \item frequency. the relative of a term in a given document
 }

\code{whales.docs.metadata} a matrix of documents metadata, where
each row that represents a document contains
 \itemize{
   \item category. the Wikipedia category assigned to an article
   \item title. the title of the Wikipedia web article
 }}
\source{
The articles are downloaded from the
 \href{http://en.wikipedia.org/wiki/Main_Page}{English Wikipedia}
 with the help of
 \href{http://www.mediawiki.org/wiki/API:Main_page}{Media Wiki}
}
\usage{
\code{data(whales.vocab)}
\code{data(whales.docs)}
\code{data(whales.docs.metadata)}
}
\description{
A subset of Wikipedia artricles under the Wikipedia category
\href{http://en.wikipedia.org/wiki/Category:Whales}{Whales}
formatted for running the LDA Gibbs sampling algorithms. This
dataset contains 153 Wikipedia articles from the following
Wikipedia subcategories.
 \itemize{
   \item Category:Baleen whales
   \item Category:Dolphins
   \item Category:Killer whales
   \item Category:Oceanic dolphins
   \item Category:Whale products
   \item Category:Whaling
 }
}
\author{
Clint P. George, April 15, 2014
}
\seealso{
\pkg{\link{lda}}

Other datasets: \code{\link{bop}}, \code{\link{bop.docs}},
  \code{\link{bop.docs.metadata}}, \code{\link{bop.vocab}};
  \code{\link{wt}}, \code{\link{wt.docs}},
  \code{\link{wt.docs.metadata}}, \code{\link{wt.vocab}}
}

